requirements.txt
#########################################################################
# # AI COVER GEN
deemix
fairseq==0.12.2
faiss-cpu==1.7.3
ffmpeg-python>=0.2.0
gradio==3.39.0
lib==4.0.0
librosa==0.9.1
numpy==1.23.5
onnxruntime_gpu
praat-parselmouth>=0.4.2
pedalboard==0.7.7
pydub==0.25.1
pyworld==0.3.4
Requests==2.31.0
scipy==1.11.1
soundfile==0.12.1
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.0.1+cu118
torchcrepe==0.0.20
tqdm==4.65.0
yt_dlp==2023.7.6
sox==1.4.1
########################################################################
# # BASE FOR FASTAPI
fastapi==0.95.2
uvicorn[standard]==0.21.1
pydantic==1.*
########################################################################
# # BASE FOR WORKER
python-dotenv
celery==5.3.1
flower==1.0.0
redis==3.5.3
Cython==0.29.21
pydub==0.25.1
gevent
boto3==1.26.148